Avoid clobbering files in other projects

+++ scripts/config.sh
@@ -1,22 +1,33 @@
 #!/usr/bin/env bash
-set -eu # exit on error or undefined variable
+set -eu                   # exit on error or undefined variable
 bash -c 'set -o pipefail' # return code of first cmd to fail in a pipeline
 
 # Env
 APP_DIR=${APP_DIR}
+GOPATH=${GOPATH}
 
+# Safety check for working on multiple projects
+if [[ "${APP_DIR}" != $(pwd) ]]; then
+  echo "To avoid clobbering files in other projects run this in APP_DIR, e.g."
+  echo "  APP_DIR=\$(pwd) ./script/config.sh"
+  exit 1
+fi
 cd ${APP_DIR}
-echo "build config cmd..."
-go build -o ${APP_DIR}/config ./cmd/config
+
+if ! test -f "${GOPATH}/bin/configu"; then
+  echo "Install https://github.com/mozey/config"
+  exit 1
+fi
 
 # Create config files if they don't exist
 if [[ ! -f ${APP_DIR}/config.dev.json ]]; then
-    echo "create dev config..."
-    cp ${APP_DIR}/sample.config.dev.json ${APP_DIR}/config.dev.json
+  echo "create dev config..."
+  cp ${APP_DIR}/sample.config.dev.json ${APP_DIR}/config.dev.json
 fi
 
 echo "generate config helper..."
 cd ${APP_DIR}
-./config -generate ./pkg/config
+${GOPATH}/bin/configu -generate ./pkg/config
 go fmt ./pkg/config/config.go
 
+echo "done `basename $0`"
